:fireworks: : Added delay for the box spawn, moved the spaan box call and fixed small typo
Now every mech will wait 2 seconds before spawning a new box
Fixed small typo ( I was missing a "this." where I was drawing the
spikes for the wheels)
Moved the spawn box call so I could make the wait function possible

+++ OOP_Assignment_2.pde
@@ -23,7 +23,7 @@ void setup()
   background(0,0,17);
   // Doing this here because the setup runs before the Map class and we need those values initialised before we create the array
   background.grid();
-  array_rows=new int[(int)background.no_boxes][(int)background.no_vert_lines]; // no_boxes and no_vert_lines are floats so we cast them into int
+  array_rows=new int[background.no_boxes][background.no_vert_lines]; 
   mech.spawn_m();
   mech.spawn_box();
